feat: steps ApiContent + Upstream components
