+++ src/components/ApiDeployModal/ApiDeployModal.tsx
@@ -1,43 +1,28 @@
-import {useEffect, useMemo, useState} from 'react';
+import {useState} from 'react';
 
-import {Button, Form, Modal, Select, Skeleton, Steps, Tabs, Tag} from 'antd';
+import {Button, Form, Modal, Steps} from 'antd';
 
 import YAML from 'yaml';
 
-import {ApiItem, ServiceItem, useDeployApi} from '@models/api';
-
-import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {setApis} from '@redux/reducers/main';
+import {useAppDispatch} from '@redux/hooks';
+// import {setApis} from '@redux/reducers/main';
 import {closeApiDeployModal} from '@redux/reducers/ui';
 
 import {ErrorLabel} from '@components/AntdCustom';
 
+import ApiContent from './ApiContent';
+import Upstream from './Upstream';
+
 import * as S from './styled';
-
-const {TabPane} = Tabs;
-
-const {Option} = Select;
 
 const ApiDeployModal: React.FC = () => {
   const dispatch = useAppDispatch();
-  const apis = useAppSelector(state => state.main.apis);
-  const services = useAppSelector(state => state.main.services);
 
-  const [activeStep, setActiveStep] = useState<number>(1);
+  const [activeStep, setActiveStep] = useState<number>(0);
   const [apiContent, setApiContent] = useState<{name: string; namespace: string; openapi: {[key: string]: any}}>();
   const [errorMessage, setErrorMessage] = useState<string>();
-  const [selectedService, setSelectedService] = useState<ServiceItem>();
-  const [upstreamReference, setUpstreamReference] = useState<string>('service');
 
-  const {mutate: deployAPI} = useDeployApi({});
-
-  const selectedServicePorts = useMemo(() => {
-    if (!selectedService) {
-      return [];
-    }
-
-    return selectedService.ports?.map(port => port.port);
-  }, [selectedService]);
+  // const {mutate: deployAPI} = useDeployApi({});
 
   const [form] = Form.useForm();
 
@@ -62,97 +47,95 @@ const ApiDeployModal: React.FC = () => {
         openapi: YAML.stringify(deployedOpenApiSpec),
       };
 
-      deployAPI(body)
-        .then((response: any) => {
-          const apiData: ApiItem = response;
+      console.log(values);
+      console.log(body);
 
-          dispatch(setApis([...apis, apiData]));
-          dispatch(closeApiDeployModal());
-        })
-        .catch(err => {
-          setErrorMessage(err.data);
-        });
+      // deployAPI(body)
+      //   .then((response: any) => {
+      //     const apiData: ApiItem = response;
+
+      //     dispatch(setApis([...apis, apiData]));
+      //     dispatch(closeApiDeployModal());
+      //   })
+      //   .catch(err => {
+      //     setErrorMessage(err.data);
+      //   });
     });
   };
 
   const onNextHandler = () => {
     form.validateFields().then(values => {
-      const {name, namespace, openapi} = values;
+      // api content
+      if (!activeStep) {
+        const {name, namespace, openapi} = values;
 
-      setApiContent({
-        name,
-        namespace: namespace || 'default',
-        openapi: YAML.parse(JSON.parse(JSON.stringify(openapi))),
-      });
-      setActiveStep(1);
-    });
-  };
+        setApiContent({
+          name,
+          namespace: namespace || 'default',
+          openapi: YAML.parse(JSON.parse(JSON.stringify(openapi))),
+        });
 
-  const onServiceSelectClearHandler = () => {
-    setSelectedService(undefined);
-  };
+        setActiveStep(activeStep + 1);
+      }
 
-  const onServiceSelectHandler = (service: ServiceItem) => {
-    setSelectedService(service);
-  };
-
-  useEffect(() => {
-    if (selectedService) {
-      form.setFieldsValue({
-        upstream: {service: {name: selectedService.name, namespace: selectedService.namespace, port: undefined}},
-      });
-
-      return;
-    }
-
-    form.resetFields();
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selectedService]);
-
-  useEffect(() => {
-    setErrorMessage('');
-
-    if (!activeStep && apiContent) {
-      form.setFieldsValue({
-        name: apiContent.name,
-        namespace: apiContent.namespace,
-        openapi: YAML.stringify(apiContent.openapi),
-      });
-    } else {
-      if (!activeStep || !apiContent) {
+      if (!apiContent) {
         return;
       }
 
-      const upstreamService = apiContent.openapi['x-kusk']?.upstream?.service;
+      // upstream
+      // if (activeStep === 1) {
+      //   const {upstream} = values;
+      //   const {pattern, substitution} = upstream.rewrite?.['rewrite_regex'];
 
-      if (upstreamService) {
-        form.setFieldsValue({
-          upstream: {
-            service: {name: upstreamService.name, namespace: upstreamService.namespace, port: upstreamService.port},
-          },
-        });
-      }
-    }
+      //   let openApiSpec = {...apiContent.openapi};
 
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeStep]);
+      //   if (pattern) {
+      //     assignNestedProperty(openApiSpec, ['x-kusk', 'upstream', 'rewrite', 'rewrite_regex', 'pattern'], pattern);
+      //   }
+
+      //   if (substitution) {
+      //     assignNestedProperty(
+      //       openApiSpec,
+      //       ['x-kusk', 'upstream', 'rewrite', 'rewrite_regex', 'substitution'],
+      //       substitution
+      //     );
+      //   }
+
+      //   console.log(values);
+
+      //   if (upstreamReference === 'service') {
+      //     openApiSpec['x-kusk'] = {...openApiSpec['x-kusk'], upstream: {...values.upstream}};
+      //   }
+      // }
+
+      setActiveStep(activeStep + 1);
+    });
+  };
+
+  const onBackHandler = () => {
+    setActiveStep(activeStep - 1);
+    setErrorMessage('');
+  };
 
   return (
     <Modal
       footer={
-        activeStep ? (
-          <>
-            <Button onClick={() => setActiveStep(0)}>Back</Button>
-            <Button type="primary" onClick={onDeployHandler}>
-              Deploy
-            </Button>
-          </>
-        ) : (
-          <Button type="primary" onClick={onNextHandler}>
-            Next
+        <>
+          {activeStep ? <Button onClick={onBackHandler}>Back</Button> : null}
+
+          <Button
+            type="primary"
+            onClick={() => {
+              if (activeStep === 5) {
+                onDeployHandler();
+              } else {
+                onNextHandler();
+              }
+            }}
+          >
+            {activeStep === 5 ? 'Deploy' : 'Next'}
           </Button>
-        )
+        </>
       }
       title="Deploy New API"
       visible
@@ -179,170 +162,9 @@ const ApiDeployModal: React.FC = () => {
             }}
           >
             {activeStep === 0 ? (
-              <>
-                <Form.Item
-                  label="Name"
-                  name="name"
-                  rules={[
-                    {required: true, message: 'Enter API name!'},
-                    {pattern: /^[a-z0-9]$|^([a-z0-9\-])*[a-z0-9]$/, message: 'Wrong pattern!'},
-                    {max: 63, type: 'string', message: 'Name is too long!'},
-                    () => {
-                      return {
-                        validator(_, value) {
-                          const namespace = form.getFieldValue('namespace') || 'default';
-
-                          if (checkDuplicateAPI(apis, `${namespace}-${value}`)) {
-                            return Promise.reject(new Error(`API name is already used in ${namespace} namespace!`));
-                          }
-
-                          return Promise.resolve();
-                        },
-                      };
-                    },
-                  ]}
-                >
-                  <S.Input placeholder="Enter API name" type="text" />
-                </Form.Item>
-
-                <Form.Item label="Namespace" name="namespace">
-                  <S.Input
-                    placeholder="Enter API namespace"
-                    type="text"
-                    onChange={() => {
-                      if (form.getFieldValue('name')) {
-                        form.validateFields(['name']);
-                      }
-                    }}
-                  />
-                </Form.Item>
-
-                <Form.Item
-                  label="OpenAPI Spec"
-                  name="openapi"
-                  rules={[
-                    {
-                      required: true,
-                      message: 'Please enter your API content!',
-                    },
-                    () => {
-                      return {
-                        validator(_, value) {
-                          if (typeof YAML.parse(JSON.parse(JSON.stringify(value))) === 'object') {
-                            return Promise.resolve();
-                          }
-
-                          return Promise.reject(new Error('Please enter a valid API content!'));
-                        },
-                      };
-                    },
-                  ]}
-                >
-                  <S.Textarea rows={10} placeholder="Enter OpenAPI Spec in YAML/JSON format" />
-                </Form.Item>
-              </>
+              <ApiContent apiContent={apiContent} form={form} />
             ) : (
-              <S.ExtensionContainer>
-                <Tabs defaultActiveKey={upstreamReference} onChange={key => setUpstreamReference(key)}>
-                  <TabPane tab="Service" key="service">
-                    {services.isLoading ? (
-                      <Skeleton.Button />
-                    ) : services.error ? (
-                      <ErrorLabel>{services.error}</ErrorLabel>
-                    ) : (
-                      <Form.Item label="Cluster service" name="service">
-                        <S.Select
-                          allowClear
-                          placeholder="Select cluster service"
-                          showSearch
-                          onClear={onServiceSelectClearHandler}
-                          onSelect={(value: any, option: any) => {
-                            onServiceSelectHandler(option.service);
-                          }}
-                        >
-                          {services.items.map(serviceItem => (
-                            <Option key={`${serviceItem.namespace}-${serviceItem.name}`} service={serviceItem}>
-                              <Tag>{serviceItem.namespace}</Tag>
-                              {serviceItem.name}
-                            </Option>
-                          ))}
-                        </S.Select>
-                      </Form.Item>
-                    )}
-
-                    <Form.Item
-                      label="Name"
-                      name={['upstream', 'service', 'name']}
-                      rules={[{required: true, message: 'Please enter name!'}]}
-                    >
-                      <S.Input disabled={Boolean(selectedService)} />
-                    </Form.Item>
-
-                    <Form.Item
-                      label="Namespace"
-                      name={['upstream', 'service', 'namespace']}
-                      rules={[{required: true, message: 'Please enter namespace!'}]}
-                    >
-                      <S.Input disabled={Boolean(selectedService)} />
-                    </Form.Item>
-
-                    <Form.Item
-                      label="Port"
-                      name={['upstream', 'service', 'port']}
-                      rules={[
-                        {required: true, message: `Please ${selectedService ? 'choose' : 'enter'} a valid port!`},
-                      ]}
-                    >
-                      {selectedService ? (
-                        <S.Select placeholder="Select port">
-                          {selectedServicePorts.map(port => (
-                            <Option key={port} value={port}>
-                              {port}
-                            </Option>
-                          ))}
-                        </S.Select>
-                      ) : (
-                        <S.Input type="number" />
-                      )}
-                    </Form.Item>
-                  </TabPane>
-
-                  <TabPane tab="Host" key="host">
-                    <Form.Item
-                      label="Hostname"
-                      name={['upstream', 'host', 'hostname']}
-                      rules={[{required: true, message: 'Please enter hostname!'}]}
-                    >
-                      <S.Input placeholder="e.g. example.com" />
-                    </Form.Item>
-
-                    <Form.Item
-                      label="Port"
-                      name={['upstream', 'host', 'port']}
-                      rules={[{required: true, message: 'Please enter port!'}]}
-                    >
-                      <S.Input type="number" />
-                    </Form.Item>
-                  </TabPane>
-                </Tabs>
-
-                <S.ExtensionSubHeading>Rewrite</S.ExtensionSubHeading>
-                <Form.Item
-                  label="Pattern"
-                  name={['upstream', 'rewrite', 'rewrite_regex', 'pattern']}
-                  rules={[{required: true, message: 'Please enter pattern!'}]}
-                >
-                  <S.Input />
-                </Form.Item>
-
-                <Form.Item
-                  label="Substitution"
-                  name={['upstream', 'rewrite', 'rewrite_regex', 'substitution']}
-                  rules={[{required: true, message: 'Please enter substitution!'}]}
-                >
-                  <S.Input />
-                </Form.Item>
-              </S.ExtensionContainer>
+              apiContent && <Upstream form={form} openApiSpec={apiContent.openapi} />
             )}
           </Form>
 
@@ -353,7 +175,16 @@ const ApiDeployModal: React.FC = () => {
   );
 };
 
-const checkDuplicateAPI = (apis: ApiItem[], apiKey: string) =>
-  apis.find(api => `${api.namespace}-${api.name}` === apiKey);
+// const assignNestedProperty = (obj: {[key: string]: any}, keyPath: string[], value: string | number | boolean) => {
+//   const lastKeyIndex = keyPath.length - 1;
+//   for (let i = 0; i < lastKeyIndex; i += 1) {
+//     const key = keyPath[i];
+//     if (!(key in obj)) {
+//       obj[key] = {};
+//     }
+//     obj = obj[key];
+//   }
+//   obj[keyPath[lastKeyIndex]] = value;
+// };
 
 export default ApiDeployModal;
