+++ src/components/ApiDeployModal/styled.tsx
@@ -22,10 +22,6 @@ export const Container = styled.div`
   }
 `;
 
-export const ExtensionContainer = styled.div`
-  margin-bottom: 10px;
-`;
-
 export const ExtensionSubHeading = styled.h4`
   font-size: 16px;
 `;
